pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "talktype"
version = "1.0.0"
description = "Push-to-talk voice typing that works everywhere"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "sounddevice>=0.4.6",
    "pynput>=1.7.6",
    "pyperclip>=1.8.2",
    "requests>=2.28.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "beaupy>=3.0.0",
]

[project.optional-dependencies]
local = ["faster-whisper>=0.10.0"]
server = ["fastapi>=0.100.0", "uvicorn>=0.22.0", "python-multipart>=0.0.6"]
all = ["talktype[local,server]"]

[project.scripts]
talktype = "talktype:main"

[tool.setuptools]
py-modules = ["talktype", "whisper_server", "setup_wizard"]
